+++ CHANGELOG.md
@@ -66,6 +66,8 @@
 - Changes to SqlServerEndpoint
   - Updated README.md with links to the examples
     ([issue #504](https://github.com/PowerShell/SqlServerDsc/issues/504)).
+- Changes to SqlServerMaxDop
+  - Minor style changes in the helper function Get-SqlDscDynamicMaxDop.
 - Changes to SqlServerMemory
   - Minor style cleanup of code in tests.
 - Changes to SqlScript